Factor out score display
